Gio's class
In my version, I tried to simplify down to one class and do things like identifying the card's name in the functions of the main program. This can, however, occur within in the class with a few tweaks. I just need to figure out how to pass that information to the class.